fix(tasks): restore the task event callback at the child turn boundary
The same stack rebase also dropped turnStep's handleEvent wiring that
forwarded a task child's input.requested / authorization.required /
authorization.completed events to the parent's session callback
(fireTaskEventCallbackStep). fireSessionCallbackStep kept its terminal
callers, but the blocking-transition forwarder lost its only call site,
so a remote task child raising HITL never reached the parent's task run:
the task stayed 'working', waitForTaskInput starved, and the remote
input-answer eval timed out.

Restore the forwarding as forwardTaskEventToSessionCallback in its own
module, and move resolveEffectiveOutputSchema (a pure helper) to
effective-output-schema.ts so workflow-steps.ts stays under the
700-line structural cap.
